python/github-sms-notify/requirements.txt
etext
PyGithub
